Smart Event Registration App Generator
Cleaver is a framework which helps developers generate an Android
application for event registrations. The goal of creating Cleaver is as
the name suggests, to cut the developers work (by a lot, hopefully) and
provide an easy way for anyone who knows Java to create an app for event
registrations. The developer does NOT need to know the first thing about
Android app development.